+++ packages/xdl/src/detach/AssetBundle.ts
@@ -11,18 +11,24 @@ import { AnyStandaloneContext } from './StandaloneContext';
 import { saveUrlToPathAsync } from './ExponentTools';
 
 const EXPO_DOMAINS = ['expo.io', 'exp.host', 'expo.test', 'localhost'];
-const ASSETS_DIR_DEFAULT_URL = 'https://d1wp6m56sqw74a.cloudfront.net/~assets';
+export const DEFAULT_CDN_HOST = 'https://d1wp6m56sqw74a.cloudfront.net';
+export const ASSETS_DIR_DEFAULT_URL = `${DEFAULT_CDN_HOST}/~assets`;
 
 type UrlResolver = (hash: string) => string;
 
-export async function bundleAsync(context: any, assets: string[], dest: string) {
+export async function bundleAsync(
+  context: any,
+  assets: string[],
+  dest: string,
+  exportUrl?: string | null
+) {
   if (!assets) {
     return;
   }
 
   await fs.ensureDir(dest);
 
-  const urlResolver = createAssetsUrlResolver(context);
+  const urlResolver = createAssetsUrlResolver(context, exportUrl);
   await pMap(assets, asset => downloadAssetAsync(urlResolver, dest, asset), { concurrency: 5 });
 }
 
@@ -33,11 +39,16 @@ async function downloadAssetAsync(urlResolver: UrlResolver, dest: string, asset:
     extensionIndex >= 0
       ? asset.substring(prefixLength, extensionIndex)
       : asset.substring(prefixLength);
+  console.log(urlResolver(hash));
   await pRetry(() => saveUrlToPathAsync(urlResolver(hash), path.join(dest, asset)), { retries: 3 });
 }
 
-function createAssetsUrlResolver(context: AnyStandaloneContext): UrlResolver {
-  let assetsDirUrl = ASSETS_DIR_DEFAULT_URL;
+function createAssetsUrlResolver(
+  context: AnyStandaloneContext,
+  exportUrl?: string | null
+): UrlResolver {
+  let assetsDirUrl = exportUrl ? `${exportUrl}/assets` : ASSETS_DIR_DEFAULT_URL;
+
   if (context && context.published && context.published.url) {
     const { assetUrlOverride = './assets' } = context.config;
     const publishedUrl = context.published.url;
@@ -52,5 +63,6 @@ function createAssetsUrlResolver(context: AnyStandaloneContext): UrlResolver {
       assetsDirUrl = url.resolve(publishedUrl, assetUrlOverride);
     }
   }
+
   return hash => urlJoin(assetsDirUrl, hash);
 }